# Conflicts:
#	src/main/app/AppBuilder.java
#	src/main/app/Main.java
#	src/main/view/InputRoomsView.java
#	src/main/view/MapPanel.java

+++ src/main/use_case/navigation/maplocation/MapLocationFactory.java
@@ -13,5 +13,5 @@ public interface MapLocationFactory {
      * @param floor the id of the floor
      * @return the created MapLocation object
      */
-    MapLocation createMapLocation(String id, double mapX, double mapY, int floor);
+    MapLocation createMapLocation(String id, int mapX, int mapY, int floor);
 }
